+++ extras/sample_site/app/models.py
@@ -7,10 +7,18 @@ class Writer(models.Model):
         return 'Writer(id=%s %s)' % (self.id, self.name)
 
 
+class Station(models.Model):
+    name = models.CharField(max_length=50)
+
+    def __str__(self):
+        return 'Station(id=%s %s)' % (self.id, self.name)
+
+
 class Show(models.Model):
     title = models.CharField(max_length=50)
     writer = models.ForeignKey(Writer, on_delete=models.CASCADE, null=True,
         blank=True)
+    stations = models.ManyToManyField(Station)
 
     def __str__(self):
         return 'Show(id=%s %s)' % (self.id, self.title)